requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.6
confluent-kafka==2.2.0
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
prometheus-client==0.17.1
psycopg2-binary==2.9.6
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
PyYAML==6.0.1
redis==4.6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
